requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
blinker==1.4
Brotli==1.0.9
certifi==2020.6.20
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dataclasses-json==0.6.7
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
duckduckgo_search==7.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
keyring==23.5.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-text-splitters==0.3.4
langsmith==0.2.6
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.3.0
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.23.2
more-itertools==8.10.0
multidict==6.1.0
mutagen==1.45.1
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
orjson==3.10.12
packaging==24.2
primp==0.9.2
propcache==0.2.1
pycairo==1.20.1
pycryptodomex==3.11.0
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pynput==1.7.7
python-apt==2.4.0+ubuntu4
python-dotenv==1.0.1
pyxattr==0.7.2
PyYAML==5.4.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.36
systemd-python==234
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
typing-inspect==0.9.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
wadllib==1.3.6
websockets==9.1
xkit==0.0.0
yarl==1.18.3
youtube-dl==2021.12.17
yt-dlp==2022.4.8
zipp==1.0.0
